% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DITHER.R
\name{DITHER}
\alias{DITHER}
\title{DITHER}
\usage{
DITHER(input_data_mut, input_data_cna)
}
\arguments{
\item{input_data_mut}{A dataframe or matrix which at least includes two columns ("Sample" and "VAF").}

\item{input_data_cna}{A dataframe or matrix which at least includes two columns ("Sample" and "Segment_Mean").}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Tumor sample ID.}
\item{DITHER_score}{The somatic mutation and CNA entropy-based ITH score in tumors.}
}
\description{
\code{DITHER} Calculate ITH based on the entropy of somatic mutation and CNA profiles in tumors.
}
\details{
This function calculates the ITH score in a bulk tumor, which is the average of the entropy of its somatic mutation profiles and the entropy of its CNA profiles.
}
\examples{
path1 = system.file("extdata", "example_mDITHER.txt", package = "DITHER", mustWork = TRUE)
path2 = system.file("extdata", "example_cDITHER.txt", package = "DITHER", mustWork = TRUE)
input_data_mut = read.table(path1, stringsAsFactors = FALSE, sep = "\\t", header = TRUE, quote = "")
input_data_cna = read.table(path2, stringsAsFactors = FALSE, sep = "\\t", header = TRUE, quote = "")
DITHER(input_data_mut, input_data_cna)
}
\author{
Lin Li <cpu_lilin@126.com>, Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
