% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDITHER.R
\name{mDITHER}
\alias{mDITHER}
\title{mDITHER}
\usage{
mDITHER(input_data_mut)
}
\arguments{
\item{input_data_mut}{A dataframe or matrix which at least includes two columns ("Sample" and "VAF").}
}
\value{
A dataframe with 2 columns:
\item{Sample}{Tumor sample ID.}
\item{mDITHER_score}{The somatic mutation entropy-based ITH score in tumors.}
}
\description{
\code{mDITHER} Calculate ITH based on the entropy of somatic mutation profiles in tumors.
}
\details{
This function calculates the ITH score in a bulk tumor based on the entropy of its somatic mutation profiles.
}
\examples{
path = system.file("extdata", "example_mDITHER.txt", package = "DITHER", mustWork = TRUE)
input_data_mut = read.table(path, stringsAsFactors = FALSE, sep = "\\t", header = TRUE, quote = "")
mDITHER(input_data_mut)
}
\author{
Lin Li <cpu_lilin@126.com>, Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
